backend/src/requirements.txt
ipykernel
pandas
cassandra-driver
openai
loguru
python-dotenv
fastapi==0.99.0
uvicorn
mangum